requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LLM and AI services
groq==0.4.1
sentence-transformers==5.0.0

# Document processing
docling==2.43.0
docling-core==2.44.1
PyPDF2==3.0.1

# Vector database
chromadb==0.4.15

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Configuration and environment
pydantic==2.7.0
pydantic-settings==2.3.1
python-dotenv==1.0.0

# Utilities
requests>=2.32.3
numpy>=1.26.4,<2.0.0
pandas>=2.0.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0

# Logging and monitoring
structlog==23.2.0